add optional test step
